refact: rename authz.counts to authz.discovery
